+++ haystack/agents/memory/__init__.py
@@ -1,4 +1,4 @@
 from haystack.agents.memory.base import Memory
-from haystack.agents.memory.base import ConversationMemory
-from haystack.agents.memory.base import NoMemory
-from haystack.agents.memory.base import ConversationSummaryMemory
+from haystack.agents.memory.no_memory import NoMemory
+from haystack.agents.memory.conversation_memory import ConversationMemory
+from haystack.agents.memory.conversation_summary_memory import ConversationSummaryMemory
